tests/wpt/tests/html/canvas/offscreen/manual/draw-generic-family/2d.text.draw.generic.family.html

<!DOCTYPE html>
<title>OffscreenCanvas test: 2d.text.draw.generic.family</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<script>
function drawCanvas(ctx, family)
{
    ctx.font = '16px ' + family;
    ctx.fillText(family, 0, 16);
}

function testDrawGenericFamily(family)
{
    let offscreenCanvas = new OffscreenCanvas(88, 24);
    let oCtx = offscreenCanvas.getContext('2d');
    drawCanvas(oCtx, family);
    let canvas = document.createElement('canvas');
    let ctx = canvas.getContext('2d');
    drawCanvas(ctx, family);

    let data1 = oCtx.getImageData(0, 0, 88, 24).data;
    let data2 = ctx.getImageData(0, 0, 88, 24).data;
    assert_array_equals(data1, data2,
      "The image data generated by drawing generic font family '" + family +
      "' should be the same for both OffscreenCanvas and regular canvas");
}

test(function() {
    testDrawGenericFamily('sans-serif');
}, "Test that drawing sans-serif produces the same result between canvas and OffscreenCanvas");

test(function() {
    testDrawGenericFamily('serif');
}, "Test that drawing serif produces the same result between canvas and OffscreenCanvas");

test(function() {
    testDrawGenericFamily('cursive');
}, "Test that drawing cursive produces the same result between canvas and OffscreenCanvas");

test(function() {
    testDrawGenericFamily('fantasy');
}, "Test that drawing fantasy produces the same result between canvas and OffscreenCanvas");

test(function() {
    testDrawGenericFamily('monospace');
}, "Test that drawing monospace produces the same result between canvas and OffscreenCanvas");
</script>
